% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainspriteR.R
\name{brainsprite}
\alias{brainsprite}
\title{Brainsprite}
\usage{
brainsprite(sprites, spriteHeight = 233, spriteWidth = 189,
  inputId = tempfile("id") \%>\% basename, spriteID = tempfile("id")
  \%>\% basename, flagCoordinates = FALSE, origin = c(0, 0, 0),
  fontColor = "#FFFFFF", background = "#000000", overlay = NULL,
  overlayHeight = NULL, overlayWidth = NULL,
  overlayID = tempfile("id") \%>\% basename, overlayOpacity = 0.5,
  height = NULL, width = NULL)
}
\arguments{
\item{sprites}{File path for the spries}

\item{spriteHeight}{Height of the brain sprite tiles}

\item{spriteWidth}{Width of the brain sprite tiles}

\item{inputId}{Input ID for use in shiny applications}

\item{spriteID}{Id for the sprite image}

\item{flagCoordinates}{Show coordinates of the clicked position}

\item{origin}{Which voxel has the coordinates 0,0,0}

\item{fontColor}{Color of the text}

\item{background}{Background color}

\item{overlay}{File path for an optional overlay image}

\item{overlayHeight}{Height of the overlay sprite tiles}

\item{overlayWidth}{Width of the overlay sprite tiles}

\item{overlayID}{Id for the overlay image}

\item{overlayOpacity}{Opacity of the overlay image}

\item{height}{Height of the visualization}

\item{width}{Width of the visualization}
}
\description{
Creates an interactive 3D viewer for MRI slices.
}
